requirements.txt

streamlit
streamlit-pydantic
pydantic

flask
gunicorn
requests
duckdb
google-auth-oauthlib
google-auth
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary